Fix compile warnings: -Wsign-compare, stage 2
All -Wsign-compare sites fixed (20 unique source locations, 76 raw
warnings across TUs). Clean build: 349 warnings left, all
-Wsign-conversion (stage 3, not started).

Two sites were the real bug this warning class exists to catch:
Player.cpp and PatternEditor.cpp both compared a signed
InstrumentTrack::getInstrumentId() against .size() with no >= 0
guard, unlike InstrumentTrackState.h's equivalent check - a negative
id would have silently promoted to a huge unsigned value and passed.
Both now guard >= 0 explicitly.

Everything else: matched the comparison's actual type at each site
(size_t loop counter vs. a single cast at the container-size bound,
whichever needed fewer total casts in context - no blanket casts).

Verified: clean rebuild (0 errors), full ctest pass, --render smoke
test.

+++ InfoLine.h
@@ -37,7 +37,7 @@ class InfoLine : public UIElement {
 
       auto s = fmt::format(" {:02x} {:02d}:{:02d} pattern:{} voices:{}/{}", info.getAbsolutePosition(), minutes, seconds, pattern_idx, num_voices, num_allocated_voices);
       if (info.isPlaying()) s += " PLAYING";
-      while (s.size() < cols) s += ' ';
+      while (s.size() < static_cast<size_t>(cols)) s += ' ';
       
       putstr(0, 0, s);
